finished

+++ app/workers/audio_converter_worker.rb
@@ -11,19 +11,21 @@ class AudioConverterWorker
 
     ext = record.file_name.split(".")[1]
 
-    command = "soundconverter -b -m audio/mpeg -s .mp3 public/uploads/" + record.unic_name + "." + ext
+    if ext != "mp3"
+      command = "soundconverter -b -m audio/mpeg -s .mp3 public/uploads/" + record.unic_name + "." + ext
+      stdin, stdout, stderr = Open3.popen3(command)
+      output = stdout.read
+      stdin.close; stdout.close; stderr.close
 
-
-    stdin, stdout, stderr = Open3.popen3(command)
-    output = stdout.read
-    stdin.close; stdout.close; stderr.close
-
-    if(output['Queue done'] == "Queue done")
-      File.unlink("public/uploads/" + record.unic_name + "." + ext)
-      file_size = File.size("public/uploads/" + record.unic_name + ".mp3")
-      record.update({:converted => 1, :size => file_size})
+      if(output['Queue done'] == "Queue done")
+        File.unlink("public/uploads/" + record.unic_name + "." + ext)
+        file_size = File.size("public/uploads/" + record.unic_name + ".mp3")
+        record.update({:converted => 1, :size => file_size})
+      else
+        record.update({:converted => 2})
+      end
     else
-      record.update({:converted => 2})
+      record.update({:converted => 1})
     end
 
   end
